+++ lib/duckdb_dart.dart
@@ -4,5 +4,6 @@ library duckdb;
 
 export 'src/ffi/duckdb.g.dart';
 export 'src/ffi/load_library.dart';
+export 'src/implementation/config.dart';
 export 'src/implementation/duckdb.dart';